Crash instance if console monitoring crashes
When the instance is started a task collecting logs is started as well.
The assumption is that if that task is correctly running the underlying
container is also running.
However if that stops unexpectedly it means that the container also
stopped.

The connection with Docker API is done over unix socket and therefore
chances are low that the connection might interrupt while the container
is still actually running.

As of now this has proven to be a sensible enough assumption for a
simple usecase.