Generate a hash for each model
We want to be sure that we are going to change the state of only those
entities that actually changed, instead of all of them.
For this reason each entity gets tagged with a hash that we use it also
in the target server. When the local hash changes then we have to
re-deploy the application.