Allow expressing the port to use when proxying the application
In order to properly load balance instances of a given application with
Traefik is better to configure the port used to expose the service.

Generally Traefik can deduct that based on different critieria but
fixing it is probably the most deterministic behaviour and it also
doesn't require the container to actively expose anything on the host
directly.